requirements.txt
# TravelBot — Requirements v5.0
# Python 3.11 | Streamlit Cloud compatible

# ── Web Framework ──────────────────────────
streamlit>=1.28.0

# ── Environment ────────────────────────────
python-dotenv>=1.0.0

# ── LLM (Groq) — Chat + Vision ─────────────
groq>=0.4.1

# ── Machine Learning ───────────────────────
scikit-learn>=1.3.0
joblib>=1.3.0

# ── NLP & Embeddings ───────────────────────
sentence-transformers>=2.2.0
nltk>=3.8

# ── Image Processing ───────────────────────
Pillow>=10.0.0

# ── Data Processing ────────────────────────
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# ── HTTP Requests (Weather API + Tour links) ─
requests>=2.31.0

# ── KHÔNG CẦN (đã bỏ) ─────────────────────
# clip / open-clip-torch  → thay bằng Groq Vision API
# torch / torchvision     → không dùng CLIP local nữa

python-docx>=1.1.0
reportlab>=4.0.0


